% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exnexstan-package.R
\docType{package}
\name{exnexstan-package}
\alias{exnexstan}
\alias{exnexstan-package}
\title{exnexstan: Fit EXNEX models with Stan}
\description{
R package for fitting Bayesian EXNEX models with Stan. Based on the original paper "Robust exchangeability designs for early phase clinical trials with multiple strata" by Neuenschwander et al.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxdrohde/exnexstan}
  \item Report bugs at \url{https://github.com/maxdrohde/exnexstan/issues}
}

}
\keyword{internal}
