% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_exch.R
\name{fit_exch}
\alias{fit_exch}
\title{Fit an exchangeable hierarchical model using a non-centered parameterization}
\usage{
fit_exch(
  n,
  r,
  seed = 123456789,
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 3000,
  iter_sampling = 5000,
  adapt_delta = 0.9,
  ...
)
}
\arguments{
\item{n}{An integer vector with the number of observations per strata}

\item{r}{An integer vector with the number of "successes" per strata}

\item{seed}{Set seed for the random number generated}

\item{chains}{Number of MCMC chains to run}

\item{parallel_chains}{Number of cores to use for running chains in parallel}

\item{iter_warmup}{Number of warmup iterations}

\item{iter_sampling}{Number of sampling iterations}

\item{adapt_delta}{Tuning parameter for MCMC sampling}

\item{...}{Other parameters to be passed into the sample function of cmdstanr}
}
\value{
A fitted cmdstanr model
}
\description{
When fitting an EXNEX model with \code{p_exch} close to 1, there can be convergence issues
due to the funnel-like geometry that arises when \eqn{\tau} is close to zero.
For the special case of \code{p_exch = 1},
this function fits the model using a non-centered parameterization,
which can fix the convergence issues.
See the \href{https://mc-stan.org/docs/stan-users-guide/reparameterization.html}{Stan user guide}
for more information.
}
\examples{
\dontrun{
# Example data from Table 1 of Neuenschwander et al.
fit_exch(r = c(2, 0, 1, 6, 7, 3, 5, 1, 0, 3) |> as.integer(),
         n = c(15, 13, 12, 28, 29, 29, 26, 5, 2, 20) |> as.integer(),
         adapt_delta = 0.99)
}

}
\seealso{
\code{\link[=fit_exnex]{fit_exnex()}}
}
