% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_exnex.R
\name{fit_exnex}
\alias{fit_exnex}
\title{Fit an EXNEX model}
\usage{
fit_exnex(
  n,
  r,
  p_exch,
  seed = 123456789,
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 3000,
  iter_sampling = 5000,
  adapt_delta = 0.9,
  ...
)
}
\arguments{
\item{n}{An integer vector with the number of observations per strata}

\item{r}{An integer vector with the number of "successes" per strata}

\item{p_exch}{A numeric vector specifying the prior on the exchangeability proportion for each strata}

\item{seed}{Set seed for the random number generated}

\item{chains}{Number of MCMC chains to run}

\item{parallel_chains}{Number of cores to use for running chains in parallel}

\item{iter_warmup}{Number of warmup iterations}

\item{iter_sampling}{Number of sampling iterations}

\item{adapt_delta}{Tuning parameter for MCMC sampling}

\item{...}{Other parameters to be passed into the `sample` function of cmdstanr}
}
\value{
A fitted cmdstanr model
}
\description{
Fit an EXNEX model
}
\examples{
\dontrun{
# Example data from Table 1 of Neuenschwander et al.
fit_exnex(r = c(2, 0, 1, 6, 7, 3, 5, 1, 0, 3) |> as.integer(),
          n = c(15, 13, 12, 28, 29, 29, 26, 5, 2, 20) |> as.integer(),
          p_exch = rep(0.5, 10),
          adapt_delta = 0.99)
}

}
