% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_exnex_varying_time.R
\name{fit_exnex_varying_time}
\alias{fit_exnex_varying_time}
\title{Fit an EXNEX model for varying follow-up time with Stan}
\usage{
fit_exnex_varying_time(
  t,
  r,
  p_exch,
  mu_prior_mean = -1.73,
  mu_prior_sd = 2.616,
  tau_lower_bound = 0,
  tau_prior_mean = 0,
  tau_prior_sd = 1,
  nex_prior_mean = -1.73,
  nex_prior_sd = 2.801,
  seed = 123456789,
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 3000,
  iter_sampling = 5000,
  adapt_delta = 0.9,
  ...
)
}
\arguments{
\item{t}{A numeric vector with the follow-up time per strata}

\item{r}{An integer vector with the number of events per strata}

\item{p_exch}{A numeric vector specifying the prior probability of exchangeability for each strata}

\item{mu_prior_mean}{Mean for the normal prior set on mu.}

\item{mu_prior_sd}{Standard deviation for the normal prior set on mu.}

\item{tau_lower_bound}{Lower bound for the tau parameter. Any probability mass below that will be reallocated.}

\item{tau_prior_mean}{Mean for the normal prior set on tau. Tau has a lower bound so any probability mass below that will be reallocated.}

\item{tau_prior_sd}{Standard deviation for the normal prior set on tau.
Tau has a lower bound of zero, so setting \code{tau_prior_sd = 1} is equivalent to a standard
half-normal distribution.}

\item{nex_prior_mean}{Mean for the normal prior set on the non-exchangeable distributions.}

\item{nex_prior_sd}{Standard deviation for the normal prior set on the non-exchangeable distributions.}

\item{seed}{Set seed for the random number generated}

\item{chains}{Number of MCMC chains to run}

\item{parallel_chains}{Number of cores to use for running chains in parallel}

\item{iter_warmup}{Number of warmup iterations}

\item{iter_sampling}{Number of sampling iterations}

\item{adapt_delta}{Tuning parameter for MCMC sampling}

\item{...}{Other parameters to be passed into the sample function of cmdstanr}
}
\value{
A fitted cmdstanr model
}
\description{
Fit an EXNEX model for varying follow-up time with Stan
}
\section{MCMC convergence issues}{

When \code{p_exch} is set close to 1, there can be convergence issues
due to the funnel-like geometry that arises when \eqn{\tau} is close to zero.
Setting the lower bound for tau can help with this.
}

\examples{
\dontrun{
fit_exnex_varying_time(
  r = rep(5, 10) |> as.integer(),
  t = seq(5, 50, length.out=10),
  p_exch = rep(0.5, 10),
  adapt_delta = 0.99
)
}

}
