% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Create a gt HTML summary table from a EXNEX object}
\usage{
summary_table(exnex_model)
}
\arguments{
\item{exnex_model}{An EXNEX model object}
}
\value{
A gt HTML table summarizing the model results
}
\description{
Create a gt HTML summary table from a EXNEX object
}
