% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_exnex.R
\name{fit_exnex}
\alias{fit_exnex}
\title{Fit an EXNEX model with Stan}
\usage{
fit_exnex(
  n,
  r,
  p_exch,
  mu_prior_mean = -1.73,
  mu_prior_sd = 2.616,
  tau_prior_mean = 0,
  tau_prior_sd = 1,
  nex_prior_mean = -1.73,
  nex_prior_sd = 2.801,
  seed = 123456789,
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 3000,
  iter_sampling = 5000,
  adapt_delta = 0.9,
  ...
)
}
\arguments{
\item{n}{An integer vector with the number of observations per strata}

\item{r}{An integer vector with the number of "successes" per strata}

\item{p_exch}{A numeric vector specifying the prior probability of exchangeability for each strata}

\item{mu_prior_mean}{Mean for the normal prior set on mu.}

\item{mu_prior_sd}{Standard deviation for the normal prior set on mu.}

\item{tau_prior_mean}{Mean for the normal prior set on tau.
Tau has a lower bound of zero so any probability mass below zero will be reallocated.
Setting \code{tau_prior_mu = 0} and \code{tau_prior_sd = 1} is equivalent to a standard
half-normal distribution.}

\item{tau_prior_sd}{Standard deviation for the normal prior set on tau.
Tau has a lower bound of zero, so setting \code{tau_prior_mu = 0} and \code{tau_prior_sd = 1} is equivalent to a standard
half-normal distribution.}

\item{nex_prior_mean}{Mean for the normal prior set on the non-exchangeable distributions.}

\item{nex_prior_sd}{Standard deviation for the normal prior set on the non-exchangeable distributions.}

\item{seed}{Set seed for the random number generated}

\item{chains}{Number of MCMC chains to run}

\item{parallel_chains}{Number of cores to use for running chains in parallel}

\item{iter_warmup}{Number of warmup iterations}

\item{iter_sampling}{Number of sampling iterations}

\item{adapt_delta}{Tuning parameter for MCMC sampling}

\item{...}{Other parameters to be passed into the sample function of cmdstanr}
}
\value{
A fitted cmdstanr model
}
\description{
Fit an EXNEX model with Stan
}
\section{MCMC convergence issues}{

When \code{p_exch} is set close to 1, there can be convergence issues
due to the funnel-like geometry that arises when \eqn{\tau} is close to zero.
See the \code{\link[=fit_exch]{fit_exch()}} function, which can fix convergence issue using
a non-centered parameterization of the model for the special case of \code{p_exch = 1}.
}

\examples{
\dontrun{
# Example data from Table 1 of Neuenschwander et al.
fit_exnex(r = c(2, 0, 1, 6, 7, 3, 5, 1, 0, 3) |> as.integer(),
          n = c(15, 13, 12, 28, 29, 29, 26, 5, 2, 20) |> as.integer(),
          p_exch = rep(0.5, 10),
          adapt_delta = 0.99)
}

}
